Rename composable mock folder
